Refactoring, warning reduction
- Add c++ latest as compile option
- Add msvc warning level W4
- Update submodule version
- Many small refactorings to reduce warnings during compilation

+++ source/hoi4/goals/Effects.cpp
@@ -45,7 +45,7 @@ void constructEffects(std::vector<EffectGrouping> &effectGroupings,
             auto max = std::stod(parts[1]);
             // try to detect if this is a floating point number or an integer
             if (min == static_cast<int>(min) && max == static_cast<int>(max)) {
-              auto selectedValue = RandNum::getRandom<int>(min, max);
+              auto selectedValue = static_cast<int>(RandNum::getRandom<double>(min, max));
               subParam = std::to_string(selectedValue);
             } else {
               auto selectedValue = RandNum::getRandom<double>(min, max);
