+++ source/vic3/Vic3Region.cpp
@@ -19,7 +19,7 @@ int Region::supportsBuilding(const BuildingType &buildingType) {
   if (resources.find(buildingType.group) != resources.end()) {
     auto &res = resources.at(buildingType.group);
     if (res.amount > 0) {
-      return res.capped ? res.amount : arableLand;
+      return static_cast<int>(res.capped ? res.amount : arableLand);
     }
   }
